[r] fix clang-tidy warnings
rewrite code to fix or suppress some clang tidy warnings.

disable checks from altera since they are mostly performance warnings,
 useful for OpenCL kernels, or highly optimized low level code,
 not a high level library such as clayknot.

disable the cyclomatic complexity check (for now at least). will
revisit this topic in the future, and will need to take a look at
false positives generated by Catch2 macros.

+++ apps/editor/src/main.cpp
@@ -129,7 +129,7 @@ auto main(int /*argc*/, char** /*argv*/) -> int
 				viewer->add_sub_viewer(&test_struct::b, "b viewer");
 				viewer->add_sub_viewer<float>(
 					[result = 0.0f](test_struct const& t) mutable -> float const* {
-						result = t.a + t.b;
+						result = static_cast<float>(t.a) + t.b;
 						return &result;
 					},
 					"a + b viewer");
@@ -238,6 +238,7 @@ auto main(int /*argc*/, char** /*argv*/) -> int
 	return 0;
 }
 
+// NOLINTNEXTLINE
 void glDebugOutput(GLenum source, GLenum type, GLuint id, GLenum severity, GLsizei /*length*/, const GLchar* message,
 	const void* /*userParam*/)
 {
